+++ js/data.js
@@ -1,0 +1,28 @@
+const DESCRIPTIONS = [
+  'Приятно, граждане, наблюдать, как стремящиеся вытеснить традиционное производство, нанотехнологии, которые представляют собой яркий пример континентально-европейского типа политической культуры, будут заблокированы в рамках своих собственных рациональных ограничений.',
+  'Кстати, действия представителей оппозиции, которые представляют собой яркий пример континентально-европейского типа политической культуры, будут призваны к ответу.',
+  'Прежде всего, понимание сути ресурсосберегающих технологий, а также свежий взгляд на привычные вещи - безусловно открывает новые горизонты для позиций, занимаемых участниками в отношении поставленных задач.',
+  'Таким образом, постоянный количественный рост и сфера нашей активности напрямую зависит от переосмысления внешнеэкономических политик.',
+  'Противоположная точка зрения подразумевает, что сделанные на базе интернет-аналитики выводы формируют глобальную экономическую сеть и при этом - своевременно верифицированы.',
+  'Равным образом, внедрение современных методик однозначно фиксирует необходимость системы массового участия.',
+];
+
+const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+const NAMES = [
+  'Егор',
+  'Света',
+  'Артем',
+  'Оксана',
+  'Николай',
+  'Екатерина',
+];
+
+export {DESCRIPTIONS, MESSAGES, NAMES};
