+++ js/utils/check-string-length.js
@@ -2,4 +2,4 @@ function checkStringLength (string, length) {
   return string.length <= length;
 }
 
-checkStringLength;
+export {checkStringLength};
